% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyr2016addins-package.R
\docType{package}
\name{nyr2016addins}
\alias{nyr2016addins}
\alias{nyr2016addins-package}
\title{NYR 2106 Addins}
\description{
Use RStudio addins to execute \R functions interactively within RStudio.
See \code{\link{selectConn}()} and \code{\link{variableProfiler}()}
for two examples of R functions that are exported.
}

